+++ changelog.md
@@ -65,6 +65,8 @@ Codion Change Log
 - SwingEntityTableModel.UpdateListener now replaces updated entities.
 - SwingEntityTableModel.replaceEntitiesByKey() now uses a key index map when looking up entities to replace, orders of magnitude faster.
 - EntityComboBoxModel.HANDLE_EDIT_EVENTS configuration value added.
+### is.codion.swing.framework.model.tools
+- DomainGenerator now uses com.squareup.javapoet to generate the source and generates a full domain implementation class.
 ### is.codion.swing.framework.ui
 - EntityEditPanel, control configuration refactored, now consistent with EntityTablePanel.
 - EntityTablePanel.PopupMenuConfig and ToolBarConfig removed, quite unnecessary.